mais melhorias

+++ backend/src/main/java/com/goodbooks/repository/BookRepository.java
@@ -1,11 +1,11 @@
 package com.goodbooks.repository;
 
-import com.goodbooks.model.Book;
+import com.goodbooks.domain.Book;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BookRepository extends JpaRepository<Book, Integer> {
+public interface BookRepository extends JpaRepository<Book, Long> {
     
 }
