+++ backend/src/main/java/com/goodbooks/repository/ReaderRepository.java
@@ -1,11 +1,11 @@
 package com.goodbooks.repository;
 
-import com.goodbooks.model.Reader;
+import com.goodbooks.domain.Reader;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ReaderRepository extends JpaRepository<Reader, Integer> {
+public interface ReaderRepository extends JpaRepository<Reader, Long> {
     
 }
